fix: fix alignment of reference item clone button
